requirements.txt
comtypes
python-docx